Auton3BallWorksbutTooSlow
added Shooter Move Med shoot for 3 ball auton and got angle and distance to work for 3 ball auton but to slow, needs to speed up things more.  the faster versions of the straight and turnSpin don't do anything.  Speed up drive straight and spin by adding tolerance to the ramp down distance, Spin was not tested